Log ActiveSupport::TaggedLogging tags as an array
Previously, DogCollar used an `ActiveSupport::TaggedLogging`
implementation that converted the log tags to a hash, where the key was
the tag converted to a symbol. These were then forwarded to Datadog as
part of the log metadata. Since the hash keys were symbols, this would
leak memory if you configured the Rails logger to log an always unique
log tag, such as the request id.

Instead of using a hash, the new implementation logs an array of log
tags.